requirements.txt
# GHOST System Requirements
# Основные зависимости для работы с Telegram и парсинга сигналов

# FastAPI для Render API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Telegram API
telethon>=1.28.0

# Async HTTP клиент для API запросов  
aiohttp>=3.8.0

# JSON и данные
pydantic>=1.10.0

# Логирование и мониторинг
python-dotenv>=0.19.0

# Natural Language Processing (для анализа новостей)
nltk>=3.8.0

# Supabase интеграция (зафиксированные совместимые версии)
supabase==2.3.4
gotrue==2.4.2
httpx==0.24.1

# SQLite для Node.js API
# (устанавливается через npm)

# Математика и анализ данных
numpy>=1.21.0

# Обработка изображений
Pillow>=9.0.0

# AI для анализа изображений и текста
openai>=1.0.0
google-generativeai>=0.3.0

# Время и даты
python-dateutil>=2.8.0

# HTTP запросы
requests>=2.28.0

# YAML конфигурация
PyYAML>=6.0.0

# Async задачи
asyncio-mqtt>=0.11.0

# Опциональные зависимости для расширенного функционала:

# Криптография (для безопасности)
cryptography>=3.4.0

# Работа с PostgreSQL (если нужно прямое подключение к Supabase)
# psycopg2-binary>=2.9.0

# Работа с SQLite (для локального кеширования)
# sqlite3 (встроен в Python)

# Парсинг HTML (если нужно скрапить веб-источники)
# beautifulsoup4>=4.11.0
# lxml>=4.9.0

# Машинное обучение (для анализа сигналов)
# scikit-learn>=1.1.0
# pandas>=1.4.0

# Графики и визуализация (для анализа)
# matplotlib>=3.5.0
# plotly>=5.10.0

# Уведомления
# plyer>=2.1.0
